requirements.txt
google-generativeai>=0.8.0
requests
python-dotenv
sendgrid
